python-mysql/requirements.txt
# ===========================================
# Python + MySQL - Basic Dependencies
# ===========================================
mysql-connector-python>=8.2.0
python-dotenv>=1.0.0

# ===========================================
# Optional - uncomment according to the project
# ===========================================

# --- Web: Flask ---
# flask>=3.0.0

# --- Web: FastAPI ---
# fastapi>=0.109.0
# uvicorn[standard]>=0.27.0

# --- Web: Django ---
# django>=5.0.0

# --- Data Science / ML ---
# pandas>=2.2.0
# numpy>=1.26.0
# scikit-learn>=1.4.0

# --- Jupyter (notebooks) ---
# jupyter>=1.0.0